requirements.txt
langchain==0.2.0
python-dotenv==1.0.1
pandas==2.2.1
aiogram==3.4.1
openai==0.28